+++ local/1_test_azure.py
@@ -2,32 +2,26 @@
 import azureml.core
 from azureml.core import Workspace
 from azureml.core import ComputeTarget, Datastore, Dataset
-# local imports
-from src.colors import Colors
-c = Colors()
 
-# TODO: Fill inn your workspace name
-workspace_name = ""
 
-# Load the workspace 
-print("INFO: Loading workspace ...")
-ws = Workspace.get(workspace_name)
-print("{}Ready to use Azure ML '{}' to work with '{}'.{}".format(c.YELLOW, azureml.core.VERSION, ws.name, c.DEFAULT))
+# Load the workspace
+ws = Workspace.get(name="") # TODO: Fill inn your workspace name
+print("Ready to use Azure ML '{}' to work with '{}'.".format(azureml.core.VERSION, ws.name))
 
 # Prints all the available compute targets
-print("{}Compute Targets:{}".format(c.GREEN, c.DEFAULT))
+print("Compute Targets:")
 for compute_name in ws.compute_targets:
     compute = ws.compute_targets[compute_name]
-    print("{}\t{}: {}{}".format(c.BOLD, compute.name, compute.type, c.DEFAULT))
+    print("\t{}: {}".format(compute.name, compute.type))
     
 # Prints all the available datastores
-print("{}Datastores:{}".format(c.GREEN, c.DEFAULT))
+print("Datastores:")
 for datastore_name in ws.datastores:
     datastore = Datastore.get(ws, datastore_name)
-    print("{}\t{}: {}{}".format(c.BOLD, datastore.name, datastore.datastore_type, c.DEFAULT))
+    print("\t{}: {}".format(datastore.name, datastore.datastore_type))
     
 # Prints all the available datasets
-print("{}Datasets:{}".format(c.GREEN, c.DEFAULT))
+print("Datasets:")
 for dataset_name in list(ws.datasets.keys()):
     dataset = Dataset.get_by_name(ws, dataset_name)
-    print("{}\t{}{}".format(c.BOLD, dataset.name, c.DEFAULT))
+    print("\t{}".format(dataset.name))
