+++ local/README.md
@@ -3,11 +3,12 @@ In this task you will learn how you can work with Azure ML from your local machi
 
 The main goal is to learn about Azure ML and the Azure ML SDK, but if would like to know more about the machinelearning algorithm and what it does, read the [Char RNN](#char-rnn) section.
 
-> :warning: It's important that you go trough the [Requirements](#requirements) section to be able to do any ot the tasks.
+> :warning: It's important that you go trough the [Requirements](#requirements) section to be able to do any of the tasks.
 
 ## Char RNN
-TODO
+The machinelearning that is used in this workshop is the Char RNN model, and is taken from [this reposetory](https://github.com/spro/char-rnn.pytorch). It is a multi-layer [Recurrent Neural Network (RNN)]() that uses the [GRU]() gating mechanism which is quite similar to the [LSTM](). 
 
+Takes in a sequence of characters and trying to predict the next character in the sequence...
 
 ## Contents
 1. [Requirements](#requirements)
@@ -16,9 +17,10 @@ TODO
 3. [Create a config](#create-a-config)
 4. [Upload dataset](#upload-dataset)
 5. [Train a model](#train-a-model)
-6. [Test trained model](#test-trained-model)
-7. [Create own scripts](#create-own-scripts)
-8. [Clean up](#clean-up)
+6. [Register a trained model](#register-a-trained-model)
+7. [Test trained model](#test-trained-model)
+8. [Create own scripts](#create-own-scripts)
+9. [Clean up](#clean-up)
 
 ## Requirements
 > :information_source: It is recomended to use a virtual envrionment for python, but it is not required.
@@ -64,7 +66,7 @@ python3 2_create_config.py
 If your script worked, you should be able to locate a `.azureml/` folder containing a `.config.json` file.
 
 #### Hints :bulb:
-- Look for **3** todos.
+- Look for **3** todos
 - Check out the documentation for the [Workspace class](https://docs.microsoft.com/en-us/python/api/overview/azure/ml/?view=azure-ml-py#workspace). 
 
 
@@ -79,44 +81,67 @@ python3 3_upload_dataset.py
 If you don't get any error messages you can go to your workspace at [https://ml.azure.com/](https://ml.azure.com/), and verify that your dataset exists in the **Datasets** tab under Assets.
 
 #### Hints :bulb:
-- Look for **2** todos.
+- Look for **3** todos
 - Check out the documentation for [register datasets](https://docs.microsoft.com/en-us/azure/machine-learning/how-to-create-register-datasets#register-datasets).
 
 ## Train a model
-In this task you will upload and run the training script(`train_char_rnn.py`). You can have a look at the script if you like, or read more about what it does in the [Char RNN](#char-rnn) section. What it essentially does is that it trains a CharRNN model on your Shakespeare dataset and generates a trained model.
+In this task you will upload and run the training script(`train_char_rnn.py`). You can have a look at the script if you like, or read more about what it does in the [Char RNN](#char-rnn) section. What it essentially does is that it trains a CharRNN model on the dataset you uploaded previously, and generates a trained model.
 
-The first thing you need to do is to set up a compute instance. If you don't have one already, you need to follow the steps in section [compute instance](https://github.com/claesgill/azure_ml_workshop/tree/issue_8_new_tasks#compute-instance) before you continue.
+The first thing you need to do is to set up a compute instance. You need to follow the steps in section [compute instance](https://github.com/claesgill/azure_ml_workshop/tree/issue_8_new_tasks#compute-instance) before you continue.
 
 Fill in the todos in `4_deploy_to_azure.py`, and run your script:
 ```sh
 python3 4_deploy_to_azure.py
 ```
 
-If your script ran successfully, go to the [https://ml.azure.com]() verify that your training script has completed, and that you have a registered model named ?? .
+If your script finish with no errors, you can monitor that your experiment run successfully in the terminal. Another option is to monitor it in your ml-workspace ([https://ml.azure.com](https://ml.azure.com). See the below steps.
+
+_NB! You may want to hit **refresh** frequently in each of the following steps_
+
+1. Navigate to the **Experiments** page and choose your experiment 
+2. Click your latest run and navigate to the **Outputs + logs** tab
+3. Expand **azureml-logs** and wait for **70_driver_log.txt** to show and click it when it does. In this log-file you watch all the outputs from the  training-script and verify that everything ran successfully
 
 #### Hints :bulb:
-- Look for **3** todos.
+- Look for **3** todos
 - Check out the documentation for [Experiments](https://docs.microsoft.com/en-us/python/api/azureml-core/azureml.core.experiment.experiment?view=azure-ml-py)
 - Learn more about [Estimators](https://docs.microsoft.com/en-us/python/api/azureml-train-core/azureml.train.estimator.estimator?view=azure-ml-py) in the documentation
 
 ## Register a trained model
-TODO
+In the previous section you may have noticed that your model was saved. However, it got saved in the container environment in Azure ML in witch don't make
+it available to us. In order to make the model available to us we can use the `Model` class, and more specifically the `register` method to upload our model to Azure. Let's go ahead and do that!
+
+Fill in the todos in `train_char_rnn.py` located in the `model` folder. Run the `4_deploy_to_azure.py` script afterwards:
+```sh
+python3 4_deploy_to_azure.py
+```
+
+To verify that your model was successfully uploaded, navigate to the **Models** page and you should see your model with the same name as you gave it.
+You can click it to see more details.
 
 #### Hints :bulb:
-- Look for **XX** todos.
-- 
+- Look for **1** todos
+- The model i saved in the outputs folder in Azure ML
+- Check out the documentation for the [Model](https://docs.microsoft.com/en-us/python/api/azureml-core/azureml.core.model.model?view=azure-ml-py) class
 
 
 ## Test trained model
+Now that you have trained and uploaded your model, it's time to test it and see how well it works. You can do this using the `5_generate.py` script.
+The only thing you'll need to do is to download the model you registered earlier using the `Model` class and `download` method.
 
-TODO
+Fill in the todo in `5_generate.py` and run it:
+```sh
+python3 5_generate.py
+```
+
+TODO - how to validate
 
 #### Hints :bulb:
-- Look for **XX** todos.
-- 
+- Look for **1** todo
+- Check out the methods in the [Model](https://docs.microsoft.com/en-us/python/api/azureml-core/azureml.core.model.model?view=azure-ml-py#methods) class documentation
+
 
 ## Create own scripts
-
 TODO
 
 #### Hints :bulb:
@@ -128,3 +153,4 @@ TODO
 
 ## TODOs
 - [ ] Add clean up script
+- [x] Rewrite train_char_rnn.py with helpers as it was before..
