+++ local/model/src/src.py
@@ -1,0 +1,102 @@
+import os
+import string
+import time
+import math
+
+from torch.autograd import Variable
+import torch.nn as nn
+import torch
+import unidecode
+
+# Reading and un-unicode-encoding data
+all_characters = string.printable
+n_characters = len(all_characters)
+
+def read_file(filename):
+    file = unidecode.unidecode(open(filename).read())
+    return file, len(file)
+
+# Turning a string into a tensor
+def char_tensor(string):
+    tensor = torch.zeros(len(string)).long()
+    for c in range(len(string)):
+        try:
+            tensor[c] = all_characters.index(string[c])
+        except:
+            continue
+    return tensor
+
+# Readable time elapsed
+def time_since(since):
+    s = time.time() - since
+    m = math.floor(s / 60)
+    s -= m * 60
+    return '%dm %ds' % (m, s)
+
+# RNN model
+class CharRNN(nn.Module):
+    def __init__(self, input_size, hidden_size, output_size, model="gru", n_layers=1):
+        super(CharRNN, self).__init__()
+        self.model = model.lower()
+        self.input_size = input_size
+        self.hidden_size = hidden_size
+        self.output_size = output_size
+        self.n_layers = n_layers
+
+        self.encoder = nn.Embedding(input_size, hidden_size)
+        if self.model == "gru":
+            self.rnn = nn.GRU(hidden_size, hidden_size, n_layers)
+        elif self.model == "lstm":
+            self.rnn = nn.LSTM(hidden_size, hidden_size, n_layers)
+        self.decoder = nn.Linear(hidden_size, output_size)
+
+    def forward(self, input, hidden):
+        batch_size = input.size(0)
+        encoded = self.encoder(input)
+        output, hidden = self.rnn(encoded.view(1, batch_size, -1), hidden)
+        output = self.decoder(output.view(batch_size, -1))
+        return output, hidden
+
+    def forward2(self, input, hidden):
+        encoded = self.encoder(input.view(1, -1))
+        output, hidden = self.rnn(encoded.view(1, 1, -1), hidden)
+        output = self.decoder(output.view(1, -1))
+        return output, hidden
+
+    def init_hidden(self, batch_size):
+        if self.model == "lstm":
+            return (Variable(torch.zeros(self.n_layers, batch_size, self.hidden_size)),
+                    Variable(torch.zeros(self.n_layers, batch_size, self.hidden_size)))
+        return Variable(torch.zeros(self.n_layers, batch_size, self.hidden_size))
+
+# Generates predicted results
+def generate(decoder, prime_str='A', predict_len=100, temperature=0.8, cuda=False):
+    hidden = decoder.init_hidden(1)
+    prime_input = Variable(char_tensor(prime_str).unsqueeze(0))
+
+    if cuda:
+        hidden = hidden.cuda()
+        prime_input = prime_input.cuda()
+    predicted = prime_str
+
+    # Use priming string to "build up" hidden state
+    for p in range(len(prime_str) - 1):
+        _, hidden = decoder(prime_input[:,p], hidden)
+        
+    inp = prime_input[:,-1]
+    
+    for p in range(predict_len):
+        output, hidden = decoder(inp, hidden)
+        
+        # Sample from the network as a multinomial distribution
+        output_dist = output.data.view(-1).div(temperature).exp()
+        top_i = torch.multinomial(output_dist, 1)[0]
+
+        # Add predicted character to string and use as next input
+        predicted_char = all_characters[top_i]
+        predicted += predicted_char
+        inp = Variable(char_tensor(predicted_char).unsqueeze(0))
+        if cuda:
+            inp = inp.cuda()
+
+    return predicted
