+++ local/old/old_model/char_rnn.py
@@ -1,0 +1,82 @@
+from torch.autograd import Variable
+import torch.nn as nn
+import torch
+import random
+
+# RNN model
+class CharRNN(nn.Module):
+    def __init__(self, input_size, hidden_size, output_size, model="gru", n_layers=1):
+        super(CharRNN, self).__init__()
+        self.model = model.lower()
+        self.input_size = input_size
+        self.hidden_size = hidden_size
+        self.output_size = output_size
+        self.n_layers = n_layers
+
+        self.encoder = nn.Embedding(input_size, hidden_size)
+        if self.model == "gru":
+            self.rnn = nn.GRU(hidden_size, hidden_size, n_layers)
+        elif self.model == "lstm":
+            self.rnn = nn.LSTM(hidden_size, hidden_size, n_layers)
+        self.decoder = nn.Linear(hidden_size, output_size)
+
+    def forward(self, input, hidden):
+        batch_size = input.size(0)
+        encoded = self.encoder(input)
+        output, hidden = self.rnn(encoded.view(1, batch_size, -1), hidden)
+        output = self.decoder(output.view(batch_size, -1))
+        return output, hidden
+
+    def forward2(self, input, hidden):
+        encoded = self.encoder(input.view(1, -1))
+        output, hidden = self.rnn(encoded.view(1, 1, -1), hidden)
+        output = self.decoder(output.view(1, -1))
+        return output, hidden
+
+    def init_hidden(self, batch_size):
+        if self.model == "lstm":
+            return (Variable(torch.zeros(self.n_layers, batch_size, self.hidden_size)),
+                    Variable(torch.zeros(self.n_layers, batch_size, self.hidden_size)))
+        return Variable(torch.zeros(self.n_layers, batch_size, self.hidden_size))
+
+def train(decoder, criterion, decoder_optimizer, inp, target, cuda, batch_size, chunk_len):
+    hidden = decoder.init_hidden(batch_size)
+    if cuda:
+        hidden = hidden.cuda()
+    decoder.zero_grad()
+    loss = 0
+
+    for c in range(chunk_len):
+        output, hidden = decoder(inp[:,c], hidden)
+        loss += criterion(output.view(batch_size, -1), target[:,c])
+
+    loss.backward()
+    decoder_optimizer.step()
+
+    return loss.item() / chunk_len
+
+# Turning a string into a tensor
+def char_tensor(string):
+    tensor = torch.zeros(len(string)).long()
+    for c in range(len(string)):
+        try:
+            tensor[c] = all_characters.index(string[c])
+        except:
+            continue
+    return tensor
+
+def random_training_set(chunk_len, batch_size, file, file_len, cuda):
+    inp = torch.LongTensor(batch_size, chunk_len)
+    target = torch.LongTensor(batch_size, chunk_len)
+    for bi in range(batch_size):
+        start_index = random.randint(0, file_len - chunk_len)
+        end_index = start_index + chunk_len + 1
+        chunk = file[start_index:end_index]
+        inp[bi] = char_tensor(chunk[:-1])
+        target[bi] = char_tensor(chunk[1:])
+    inp = Variable(inp)
+    target = Variable(target)
+    if cuda:
+        inp = inp.cuda()
+        target = target.cuda()
+    return inp, target
