+++ local/old/old_model/generate.py
@@ -1,0 +1,48 @@
+from torch.autograd import Variable
+import torch
+import string
+
+
+# Generates predicted results
+def generate(decoder, prime_str='A', predict_len=100, temperature=0.8, cuda=False):
+    all_characters = string.printable
+    hidden = decoder.init_hidden(1)
+    prime_input = Variable(char_tensor(prime_str).unsqueeze(0))
+
+    if cuda:
+        hidden = hidden.cuda()
+        prime_input = prime_input.cuda()
+    predicted = prime_str
+
+    # Use priming string to "build up" hidden state
+    for p in range(len(prime_str) - 1):
+        _, hidden = decoder(prime_input[:,p], hidden)
+        
+    inp = prime_input[:,-1]
+    
+    for p in range(predict_len):
+        output, hidden = decoder(inp, hidden)
+        
+        # Sample from the network as a multinomial distribution
+        output_dist = output.data.view(-1).div(temperature).exp()
+        top_i = torch.multinomial(output_dist, 1)[0]
+
+        # Add predicted character to string and use as next input
+        predicted_char = all_characters[top_i]
+        predicted += predicted_char
+        inp = Variable(char_tensor(predicted_char).unsqueeze(0))
+        if cuda:
+            inp = inp.cuda()
+
+    return predicted
+
+# Turning a string into a tensor
+def char_tensor(inp_string):
+    all_characters = string.printable
+    tensor = torch.zeros(len(inp_string)).long()
+    for c in range(len(inp_string)):
+        try:
+            tensor[c] = all_characters.index(inp_string[c])
+        except:
+            continue
+    return tensor
