--- local/train_char_rnn.py
@@ -1,211 +1,0 @@
-import argparse
-import os
-import string
-import random
-import time
-import math
-
-from azureml.core import Workspace, Dataset, Run, Experiment, RunConfiguration, ScriptRunConfig
-from torch.autograd import Variable
-import torch.nn as nn
-import torch
-import unidecode
-from tqdm import tqdm
-
-# Get experiment run context
-run = Run.get_context()
-
-# Reading and un-unicode-encoding data
-all_characters = string.printable
-n_characters = len(all_characters)
-
-def read_file(filename):
-    file = unidecode.unidecode(open(filename).read())
-    return file, len(file)
-
-# Turning a string into a tensor
-def char_tensor(string):
-    tensor = torch.zeros(len(string)).long()
-    for c in range(len(string)):
-        try:
-            tensor[c] = all_characters.index(string[c])
-        except:
-            continue
-    return tensor
-
-# Readable time elapsed
-def time_since(since):
-    s = time.time() - since
-    m = math.floor(s / 60)
-    s -= m * 60
-    return '%dm %ds' % (m, s)
-
-# RNN model
-class CharRNN(nn.Module):
-    def __init__(self, input_size, hidden_size, output_size, model="gru", n_layers=1):
-        super(CharRNN, self).__init__()
-        self.model = model.lower()
-        self.input_size = input_size
-        self.hidden_size = hidden_size
-        self.output_size = output_size
-        self.n_layers = n_layers
-
-        self.encoder = nn.Embedding(input_size, hidden_size)
-        if self.model == "gru":
-            self.rnn = nn.GRU(hidden_size, hidden_size, n_layers)
-        elif self.model == "lstm":
-            self.rnn = nn.LSTM(hidden_size, hidden_size, n_layers)
-        self.decoder = nn.Linear(hidden_size, output_size)
-
-    def forward(self, input, hidden):
-        batch_size = input.size(0)
-        encoded = self.encoder(input)
-        output, hidden = self.rnn(encoded.view(1, batch_size, -1), hidden)
-        output = self.decoder(output.view(batch_size, -1))
-        return output, hidden
-
-    def forward2(self, input, hidden):
-        encoded = self.encoder(input.view(1, -1))
-        output, hidden = self.rnn(encoded.view(1, 1, -1), hidden)
-        output = self.decoder(output.view(1, -1))
-        return output, hidden
-
-    def init_hidden(self, batch_size):
-        if self.model == "lstm":
-            return (Variable(torch.zeros(self.n_layers, batch_size, self.hidden_size)),
-                    Variable(torch.zeros(self.n_layers, batch_size, self.hidden_size)))
-        return Variable(torch.zeros(self.n_layers, batch_size, self.hidden_size))
-
-# Generates predicted results
-def generate(decoder, prime_str='A', predict_len=100, temperature=0.8, cuda=False):
-    hidden = decoder.init_hidden(1)
-    prime_input = Variable(char_tensor(prime_str).unsqueeze(0))
-
-    if cuda:
-        hidden = hidden.cuda()
-        prime_input = prime_input.cuda()
-    predicted = prime_str
-
-    # Use priming string to "build up" hidden state
-    for p in range(len(prime_str) - 1):
-        _, hidden = decoder(prime_input[:,p], hidden)
-        
-    inp = prime_input[:,-1]
-    
-    for p in range(predict_len):
-        output, hidden = decoder(inp, hidden)
-        
-        # Sample from the network as a multinomial distribution
-        output_dist = output.data.view(-1).div(temperature).exp()
-        top_i = torch.multinomial(output_dist, 1)[0]
-
-        # Add predicted character to string and use as next input
-        predicted_char = all_characters[top_i]
-        predicted += predicted_char
-        inp = Variable(char_tensor(predicted_char).unsqueeze(0))
-        if cuda:
-            inp = inp.cuda()
-
-    return predicted
-
-# Parse command line arguments
-argparser = argparse.ArgumentParser()
-argparser.add_argument('--filename',      type=str, default='shakespeare.txt')
-argparser.add_argument('--model',         type=str, default='gru')
-argparser.add_argument('--n_epochs',      type=int, default=10)
-argparser.add_argument('--print_every',   type=int, default=100)
-argparser.add_argument('--hidden_size',   type=int, default=100)
-argparser.add_argument('--n_layers',      type=int, default=2)
-argparser.add_argument('--learning_rate', type=float, default=0.01)
-argparser.add_argument('--chunk_len',     type=int, default=200)
-argparser.add_argument('--batch_size',    type=int, default=100)
-argparser.add_argument('--shuffle',       action='store_true')
-argparser.add_argument('--cuda',          action='store_true')
-args = argparser.parse_args()
-
-if args.cuda: print("Using CUDA")
-
-# Get a dataset by name
-# Downloads dataset from azure
-ws = Workspace.get('ml-aisl')
-dataset = Dataset.get_by_name(ws, name='shakespare_1')
-dataset.download(target_path='.', overwrite=True)
-file, file_len = read_file('shakespare.txt')
-
-def random_training_set(chunk_len, batch_size):
-    inp = torch.LongTensor(batch_size, chunk_len)
-    target = torch.LongTensor(batch_size, chunk_len)
-    for bi in range(batch_size):
-        start_index = random.randint(0, file_len - chunk_len)
-        end_index = start_index + chunk_len + 1
-        chunk = file[start_index:end_index]
-        inp[bi] = char_tensor(chunk[:-1])
-        target[bi] = char_tensor(chunk[1:])
-    inp = Variable(inp)
-    target = Variable(target)
-    if args.cuda:
-        inp = inp.cuda()
-        target = target.cuda()
-    return inp, target
-
-def train(inp, target):
-    hidden = decoder.init_hidden(args.batch_size)
-    if args.cuda:
-        hidden = hidden.cuda()
-    decoder.zero_grad()
-    loss = 0
-
-    for c in range(args.chunk_len):
-        output, hidden = decoder(inp[:,c], hidden)
-        loss += criterion(output.view(args.batch_size, -1), target[:,c])
-
-    loss.backward()
-    decoder_optimizer.step()
-
-    # return loss.data[0] / args.chunk_len
-    return loss.data / args.chunk_len
-
-def save():
-    save_filename = os.path.splitext(os.path.basename(args.filename))[0] + '.pt'
-    torch.save(decoder, save_filename)
-    print('Saved as {}'.format(save_filename))
-
-# Initialize models and start training
-decoder = CharRNN(
-    n_characters,
-    args.hidden_size,
-    n_characters,
-    model=args.model,
-    n_layers=args.n_layers,
-)
-decoder_optimizer = torch.optim.Adam(decoder.parameters(), lr=args.learning_rate)
-criterion = nn.CrossEntropyLoss()
-
-if args.cuda:
-    decoder.cuda()
-
-start = time.time()
-all_losses = []
-loss_avg = 0
-
-try:
-    print("Training for %d epochs..." % args.n_epochs)
-    for epoch in tqdm(range(1, args.n_epochs + 1)):
-        loss = train(*random_training_set(args.chunk_len, args.batch_size))
-        loss_avg += loss
-
-        if epoch % args.print_every == 0:
-            print('[%s (%d %d%%) %.4f]' % (time_since(start), epoch, epoch / args.n_epochs * 100, loss))
-            print(generate(decoder, 'Wh', 100, cuda=args.cuda), '\n')
-
-    print("Saving...")
-    save()
-
-except KeyboardInterrupt:
-    print("Saving before quit...")
-    save()
-
-
-
-# Complete the run
-run.complete()
